+++ w1_py_parse/schemas/__init__.py
@@ -3,6 +3,7 @@ from .da_permit import DA_PERMIT_FIELDS
 from .da_field import DA_FIELD_FIELDS
 from .da_field_specific import DA_FIELD_SPECIFIC_FIELDS
 from .da_field_bhl import DA_FIELD_BHL_FIELDS
+from .da_can_restr import DA_CAN_RESTR_FIELDS
 
 SCHEMA_ID_TO_NAME = {
     "01": "DAROOT",
@@ -31,4 +32,5 @@ SCHEMA_FIELDS = {
     "03": DA_FIELD_FIELDS,
     "04": DA_FIELD_SPECIFIC_FIELDS,
     "05": DA_FIELD_BHL_FIELDS,
+    "06": DA_CAN_RESTR_FIELDS,
 }